clean up locale code
1) 'app_lang' in config is not needed. Instead we the first element of
'locales' is the default language. If there is no 1st element then the
default is 'en'.
i18n is now disabled whenever 'locales' is none or has less than 2
elements. (previously case of 1 element was unclear)
The idea of using a null locale when i18n is disabled is removed - there
is always a locale. There has to be a default.
2) move LocaleStrings to i18n module.
3) replace BaseHandler.locale member (holding the current locale as a
tag) with BaseHandler.localeStrings (holding a LocaleStrings)
4) remove redundant code